Respect min interval returned by the tracker for failed announces
From: https://github.com/rakshasa/libtorrent/pull/246

Fixes: https://github.com/rakshasa/rtorrent/issues/1215

This changes the default values for the base class to 0, and relies on
the successful parsing of an announce to set them to something
sane. If the interval can't be parsed during a failure (from an
unrelated failure or just the keys missing), it fails back to the
original backoff behavior. However, in order to accommodate the
previous logic without modifying the ABI, the backoff timeout is
capped at 299 seconds instead of 320.